# docs/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: TODO-API
host: todo-7.herokuapp.com
basePath: /api/v1/
schemes:
#- http
- https
consumes:
- application/json
produces:
- application/json
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      description: Consumes email and sends an api key to the end-user
      operationId: get-api-key
      parameters:
        - name: body
          in: body
          required: true
          schema: 
            example: 
              email: "16211a05f7@bvrit.ac.in"
              
      responses:
        "201":
          description: Created. A verification link has been sent to your email, which is valid for an hour, please verify your email in order to login.
        "400":
          description: Bad request, caused by passing invalid/null values for email and password field.
        "500":
          description: Internal server error.  
      security:
        - JWT: []      
    x-swagger-router-controller: REGISTER
  /auth/google:
    post:
      tags:
        - auth
      description: Consumes id-token and sends an user details to the end-user
      operationId: google-signin
      parameters:
        - name: body
          in: body
          required: true
          schema: 
            example: 
              id_token: "<token here>"
              
      responses:
        "200":
          description: Successfully loggedin using Google.        
        "201":
          description: Successfully registered using Google.
        "400":
          description: Bad request, caused by passing invalid/null values for email and password field.
        "500":
          description: Internal server error.  
      security:
        - JWT: []      
    x-swagger-router-controller: REGISTER    
  /auth/token:
    post:
      tags:
        - auth
      description: Consumes refresh-token and creates a new access-token 
      operationId: new-access-token
      parameters:
        - name: header
          in: header
          required: true
          schema: 
            example: 
              refresh-token: "<token here>"
              
      responses:
        "200":
          description: Successfully created new access token  
        "400":
          description: Bad request, caused by passing invalid/null values for token.
        "500":
          description: Internal server error.  
      security:
        - JWT: []      
    x-swagger-router-controller: REGISTER    
...